change tools

+++ tools/train_jagger.sh
@@ -1,14 +1,25 @@
 #!/bin/bash
+WORK_DIR=$(pwd)
 cd $(dirname $0)
 
-IPADIC_DIR=${1:-../work/mecab-ipadic-2.7.0-20070801}
+if [ -z "$1" ]; then
+    MECAB_DIC_DIR=../work/mecab-ipadic-2.7.0-20070801
+else
+    MECAB_DIC_DIR=${WORK_DIR}/$1
+fi
+
+if [ ! -d ${MECAB_DIC_DIR} ]; then
+    echo "${MECAB_DIC_DIR} not existed."
+    exit 1
+fi
+
 MODEL=${2:-../model/jagger-ipadic}
 SPLIT_DIR=../work/split
 mkdir -p $MODEL
 rm $MODEL/patterns*
 
 echo "Training"
-train_jagger -d <(cat $IPADIC_DIR/*.csv) ../work/train.mecab > $MODEL/patterns
+train_jagger -d <(cat $MECAB_DIC_DIR/*.csv) ../work/train.mecab > $MODEL/patterns
 
 echo "Evaluation"
 cat $SPLIT_DIR/preprocess.test.*.mecab > ../work/preprocess.test.mecab
